+++ Modules/Microsoft365DSC/Examples/Resources/IntuneAccountProtectionLocalUserGroupMembershipPolicy/2-Update.ps1
@@ -28,12 +28,12 @@ Configuration Example
             Description              = "My revised description";
             Ensure                   = "Present";
             Assignments              = @(); # Updated Property
-            LocalUserGroupCollection = @(
-                MSFT_IntuneAccountProtectionLocalUserGroupCollection{
-                    LocalGroups = @('administrators', 'users')
-                    Members = @('S-1-12-1-1167842105-1150511762-402702254-1917434032')
-                    Action = 'add_update'
-                    UserSelectionType = 'users'
+            AccessGroup = @(
+                MSFT_MicrosoftGraphIntuneSettingsCatalogAccessGroup{
+                    desc = @('administrators', 'users')
+                    member = @('S-1-12-1-1167842105-1150511762-402702254-1917434032')
+                    action = 'add_update'
+                    userselectiontype = 'users'
                 }
             );
             ApplicationId         = $ApplicationId;