+++ Tests/Unit/Microsoft365DSC/Microsoft365DSC.IntuneAppAndBrowserIsolationPolicyWindows10.Tests.ps1
@@ -71,6 +71,12 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                                 OffsetUri = '/InstallWindowsDefenderApplicationGuard'
                                 AdditionalProperties = @{
                                     '@odata.type' = '#microsoft.graph.deviceManagementConfigurationChoiceSettingDefinition'
+                                    options = @(
+                                        @{
+                                            itemId = 'device_vendor_msft_windowsdefenderapplicationguard_installwindowsdefenderapplicationguard_install'
+                                            name = 'Install'
+                                        }
+                                    )
                                 }
                             },
                             @{
@@ -79,6 +85,12 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                                 OffsetUri = '/Settings/AllowWindowsDefenderApplicationGuard'
                                 AdditionalProperties = @{
                                     '@odata.type' = '#microsoft.graph.deviceManagementConfigurationChoiceSettingDefinition'
+                                    options = @(
+                                        @{
+                                            itemId = 'device_vendor_msft_windowsdefenderapplicationguard_settings_allowwindowsdefenderapplicationguard_1'
+                                            name = 'Enable Microsoft Defender Application Guard for Microsoft Edge ONLY'
+                                        }
+                                    )
                                 }
                             },
                             @{
@@ -87,6 +99,12 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                                 OffsetUri = '/Settings/AllowPersistence'
                                 AdditionalProperties = @{
                                     '@odata.type' = '#microsoft.graph.deviceManagementConfigurationChoiceSettingDefinition'
+                                    options = @(
+                                        @{
+                                            itemId = 'device_vendor_msft_windowsdefenderapplicationguard_settings_allowpersistence_0'
+                                            name = 'Application Guard discards user-downloaded files and other items (such as, cookies, Favorites, and so on) during machine restart or user log-off.'
+                                        }
+                                    )
                                 }
                             },
                             @{
@@ -95,6 +113,12 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                                 OffsetUri = '/Settings/AllowVirtualGPU'
                                 AdditionalProperties = @{
                                     '@odata.type' = '#microsoft.graph.deviceManagementConfigurationChoiceSettingDefinition'
+                                    options = @(
+                                        @{
+                                            itemId = 'device_vendor_msft_windowsdefenderapplicationguard_settings_allowvirtualgpu_0'
+                                            name = 'Cannot access the vGPU and uses the CPU to support rendering graphics. When the policy is not configured, it is the same as disabled (0).'
+                                        }
+                                    )
                                 }
                             },
                             @{
@@ -103,6 +127,12 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                                 OffsetUri = '/Settings/AllowCameraMicrophoneRedirection'
                                 AdditionalProperties = @{
                                     '@odata.type' = '#microsoft.graph.deviceManagementConfigurationChoiceSettingDefinition'
+                                    options = @(
+                                        @{
+                                            itemId = 'device_vendor_msft_windowsdefenderapplicationguard_settings_allowcameramicrophoneredirection_1'
+                                            name = 'Turns on the functionality to allow Microsoft Defender Application Guard to access the device''s camera and microphone.'
+                                        }
+                                    )
                                 }
                             }
                         )