+++ Tests/Unit/Microsoft365DSC/Microsoft365DSC.IntuneAppAndBrowserIsolationPolicyWindows10ConfigMgr.Tests.ps1
@@ -70,6 +70,12 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                                 Name = 'InstallWindowsDefenderApplicationGuard'
                                 AdditionalProperties = @{
                                     '@odata.type' = '#microsoft.graph.deviceManagementConfigurationChoiceSettingDefinition'
+                                    options = @(
+                                        @{
+                                            itemId = 'device_vendor_msft_windowsdefenderapplicationguard_installwindowsdefenderapplicationguard_install'
+                                            name = 'Install'
+                                        }
+                                    )
                                 }
                             },
                             @{
@@ -77,6 +83,12 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                                 Name = 'AllowWindowsDefenderApplicationGuard'
                                 AdditionalProperties = @{
                                     '@odata.type' = '#microsoft.graph.deviceManagementConfigurationChoiceSettingDefinition'
+                                    options = @(
+                                        @{
+                                            itemId = 'device_vendor_msft_windowsdefenderapplicationguard_settings_allowwindowsdefenderapplicationguard_1'
+                                            name = 'Enable Microsoft Defender Application Guard for Microsoft Edge ONLY'
+                                        }
+                                    )
                                 }
                             },
                             @{
@@ -84,6 +96,12 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                                 Name = 'AllowPersistence'
                                 AdditionalProperties = @{
                                     '@odata.type' = '#microsoft.graph.deviceManagementConfigurationChoiceSettingDefinition'
+                                    options = @(
+                                        @{
+                                            itemId = 'device_vendor_msft_windowsdefenderapplicationguard_settings_allowpersistence_0'
+                                            name = 'Application Guard discards user-downloaded files and other items (such as, cookies, Favorites, and so on) during machine restart or user log-off.'
+                                        }
+                                    )
                                 }
                             },
                             @{
@@ -91,6 +109,12 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                                 Name = 'AllowVirtualGPU'
                                 AdditionalProperties = @{
                                     '@odata.type' = '#microsoft.graph.deviceManagementConfigurationChoiceSettingDefinition'
+                                    options = @(
+                                        @{
+                                            itemId = 'device_vendor_msft_windowsdefenderapplicationguard_settings_allowvirtualgpu_0'
+                                            name = 'Cannot access the vGPU and uses the CPU to support rendering graphics. When the policy is not configured, it is the same as disabled (0).'
+                                        }
+                                    )
                                 }
                             },
                             @{
@@ -98,6 +122,12 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                                 Name = 'AllowCameraMicrophoneRedirection'
                                 AdditionalProperties = @{
                                     '@odata.type' = '#microsoft.graph.deviceManagementConfigurationChoiceSettingDefinition'
+                                    options = @(
+                                        @{
+                                            itemId = 'device_vendor_msft_windowsdefenderapplicationguard_settings_allowcameramicrophoneredirection_1'
+                                            name = 'Turns on the functionality to allow Microsoft Defender Application Guard to access the device''s camera and microphone.'
+                                        }
+                                    )
                                 }
                             }
                         )